chore(android): rename application package
