+++ clients/android/app/src/main/java/ai/vellum/assistant/ConnectDeepLink.java
@@ -1,4 +1,4 @@
-package ai.vocify.vellumassistant;
+package ai.vellum.assistant;
 
 import java.io.ByteArrayOutputStream;
 import java.net.URI;
