+++ clients/android/app/src/main/java/ai/vellum/assistant/NativeBiometricPlugin.java
@@ -1,4 +1,4 @@
-package ai.vocify.vellumassistant;
+package ai.vellum.assistant;
 
 import android.app.Activity;
 import android.security.keystore.KeyPermanentlyInvalidatedException;
