+++ clients/android/app/src/main/java/ai/vellum/assistant/VoiceAudioSessionPlugin.java
@@ -1,4 +1,4 @@
-package ai.vocify.vellumassistant;
+package ai.vellum.assistant;
 
 import android.content.Context;
 import android.media.AudioAttributes;
