+++ clients/ios/README.md
@@ -308,7 +308,7 @@ There's a deliberate mismatch:
 | Where                       | Value                                      |
 | --------------------------- | ------------------------------------------ |
 | Xcode `PRODUCT_BUNDLE_IDENTIFIER` | `ai.vocify-inc.vellum-assistant-ios` ← real one |
-| `capacitor.config.ts` `appId`     | `ai.vocify.vellumassistant`                     |
+| `capacitor.config.ts` `appId`     | `ai.vellum.assistant`                     |
 
 Capacitor requires `appId` to use Java package form because Android uses it as
 the application namespace. The iOS project is generated from `project.yml`
